+++ changelog.md
@@ -1,5 +1,12 @@
 ## Changelog for TDA API Tool
 
+## [0.3.0]
+### [No-Tracker] - 2020-08-04
+- Bugfix for single option-chain method call
+- Added argparse with various arguments and defaults
+- Added Account feature to collect current held equities and options
+- Added display to show held equities and options
+
 ## [0.2.0]
 ### [No-Tracker] - 2020-08-02
 - Moved account management to own file